# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Example of a user API with Swagger.",
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/last-weights/:id": {
            "get": {
                "description": "get last six weights horse and date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weights"
                ],
                "summary": "get last six weights horse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Horse Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weight add",
                        "schema": {
                            "$ref": "#/definitions/model.Weights"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/weight/:id": {
            "get": {
                "description": "get last weight horse and date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weights"
                ],
                "summary": "get last a weight horse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Horse Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weight add",
                        "schema": {
                            "$ref": "#/definitions/model.Weights"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "add a weight horse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weights"
                ],
                "summary": "add a weight horse",
                "parameters": [
                    {
                        "description": "Weight, fk_horse_id",
                        "name": "addhorse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Weights"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Weight add",
                        "schema": {
                            "$ref": "#/definitions/model.Weights"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/weights/:id": {
            "get": {
                "description": "get weights horse and date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weights"
                ],
                "summary": "get weights horse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Horse Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weight add",
                        "schema": {
                            "$ref": "#/definitions/model.Weights"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Weights": {
            "type": "object",
            "required": [
                "date",
                "weight"
            ],
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "FkHorseId": {
                    "type": "integer"
                },
                "UpdateddAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "diiferenceWeight": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastWeight": {
                    "type": "integer"
                },
                "last_date": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        }
    }
}